Figured out how to move it!
Works beautifully now. Only issue is that there’s some mutation going
on in the code -_- But considering the fundamental operation is
mutating the position of the button, might not really be a way around
that. I could probably clean the code up by moving where I store said
values that are being mutated, but it’d be pretty hard to get rid of it
all together. hmm…