Figured out how to stop it from bailing out
Turned out to be a pretty easy solution - the option was configured on
in mocha.opts in the first place. I might add it back in later once
more of the tests are working, but I wanted to be able to see that the
test code I wrote was accessing the object values correctly. Which it
is now.

On to the source code!