Finished! Woohoo!
The circles wrap as soon as the position is outside of the canvas
bounds (which is the exact center of the circles). It would look nicer
if we were wrapping only the parts of the circle that were outside
bound, but that’s a bit beyond the scope of this exercise :p.