Looks perfect when body isn't cc'd up
So, if I move the screen small enough so that the canvas gets crammed
to the left, or if I remove the body styling, it looks pretty rock
solid. BUT… not on fullscreen. Gonna look into different ways to get
the canvas’s offset from the page - bounding box method??