Added stats collection, some minor cleanups
Removed/renamed/factored out a couple of constants in addition to
collecting a bunch of stats based off of what the player does. In
particular, I keep track of times the game pauses (from mouse leaving
bounds of canvas), number of times the player tries to fire but can’t
(since the ship is ‘still charging’) and the number of times the player
successfully shoots (e.g. total times clicked - number clicks made when
ship was ‘still charging’)