Fixed ship firing not resetting correctly
It was sometimes taking a while longer than it should have to allow us
to fire again. Turned out to be a bug due to always setting the new
laser position’s ‘charging laser’ flag when moving the laser, rather
than setting it to be what the old laser’s was. This meant that a
‘dead’ laser was potentially still listed as ‘charging’ after looping
around - now that’s been fixed.