Stats/cookies working perfectly now
Stuff that I want to carries over across multiple games (just the ‘most
rounds won’ which is a best-ever thing) and the stuff I want to just
carry over for one game (the other stats) carries over successfully
just the once (such as last game’s score - not the best score overall -
and the other statistics such as fired lasers, fired lasers while the
ship was still charging, etc).

Although I don’t have a ‘game over’ message atm, if the cats get past
your ship it’s impossible to move on to the next level - meaning you
can’t continue to play and you r score cannot increase. Which means the
only way forward is to refresh the page, as I want.

Note that I purposefully coded the game to allow for basically
unlimited rounds, each one getting harder and harder, in order to keep
the true space invaders spirit. However, at some point - probably not
much more than round 10 - it will be practically impossible to beat the
round because the cats will be moving so quickly. So that puts a
practical time limit on the game of under 10 minutes).