Everything's working!
A bit ugly, but correct. Took out some of the console logs to make the
server less ‘noisy.’ The way I made username/password work is that
it’ll return something like {username:’USERNAMEINPUTTED’,
result:’failure’} if we can find a username field and a non-empty
username, regardless of if we can find password or even find a password
field. In other words I consider those a valid request, just that it
failed to log in (even if the password field is a bit mangled).
However, if we can’t even find a username, then I just return the
default {}.

I do something similar with the logout.

Not 100% sure that 402 is really what I should be sending back, but I
think that’s what Dr. Pollack said in class?