I think I see the bug now
Found the issue. Now articles and articleIDs are getting updated
correctly :)

Only things to change in the future is how IDs are generated (check for
duplicate number IDs? or just ensure the starting IDs are always their
index) and how date is stored (want same format as others).

That being said, I’m not sure why the page is not updating to show the
new articles - maybe gotta call something so it knows the global state
has changed?