Help from the professor
Dr. Pollack helped me deal with those weird import/mocking errors that
broke all my stuff. Woo! Note that we gave up on the mocha test thing
because it stayed broken even with an empty file in a subdirectory -
just ended up moving it up a level into the top level of src.