Working, with one little snag:
so long as the curl has  -H "Content-Type: application/json" it’s
perfect, however I’m not sure how to get it to parse the body if it
doesn’t explicitly set it to json though.

For the example curl requests dr. Pollack provided on piazza I wasn’t
able to access the POST payloads at a - req.body and req.query returned
{}, req.text returned undefined :’(.