Working on article image
Gotta go now, but making very good progress. Gotta check whether to
dispatch an article add action or an error message action depending on
if image is set but text isn’t. Also gotta go deal with the buttons for
text - there’s not a way to clear your text by using a button, and the
button itself isn’t a ‘reset’ button. In fact, I should go double check
all my entry/button areas have the proper stuff I want (a ‘clear the
text/file’ button of type=‘reset’ and a ‘submit’ button which may or
may not be of type reset).