Updated results
(Now I’m including results of the tests with the profile ones added in).

To any graders - you will notice copious use of findNames throughout my
testCode. I purposefully gave name values to various components rather
than IDs ONLY for the purpose of testing. Since I didn’t want to
potentially interfere with the usage of actual code, I decided not to
use IDs for the test purpose (in other words, whenever in the fronted
you see a name={blah},  you can think of that as a hook for selenium to
use).