AHA! Dr. Pollack showed me why it was locally not working (because of the whole process.env stuff)
So, it turns out that doing process.env is something form node - and on
heroku it sets process.env.REDIS_URL. So the REDIS_URL we were defining
above was actually never being used! There wasn’t any weird syntax
thingy going on at all, it was a simple case of accessing something
from inside an object (which is again null on our local machine).