Now password change is disabled when logged into Facebook
Even though it *shouldn’t* cause any issues to do a password change on
a Facebook person (would actually be pretty close to the account merge
feature) now that I’ve fixed a couple of other things, this change
isn’t that important. But regardless, I figure it’s best to completely
rule out the possibility.